enhancement: add swagger documentation

+++ src/main/java/com/corentinesgi/booksmaker/controller/BookController.java
@@ -17,6 +17,12 @@ import org.springframework.web.bind.annotation.RestController;
 import com.corentinesgi.booksmaker.model.Book;
 import com.corentinesgi.booksmaker.service.BookService;
 
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+
 @RestController
 @RequestMapping("/api/books")
 public class BookController {
@@ -26,31 +32,76 @@ public class BookController {
     this.service = service;
   }
 
+  @ApiResponses(value = {
+    @ApiResponse(
+      responseCode = "200",
+      description = "Livres récupérés avec succès",
+      content = @Content(array = @ArraySchema(schema = @Schema(implementation = Book.class)))
+    ),
+  })
   @GetMapping
   public ResponseEntity<List<Book>> getAll() {
     List<Book> results = service.findAll();
-    return ResponseEntity.status(HttpStatus.CREATED).body(results);
+    return ResponseEntity.status(HttpStatus.OK).body(results);
   }
 
+  @ApiResponses(value = {
+    @ApiResponse(
+      responseCode = "201",
+      description = "Livre créé avec succès",
+      content = @Content(schema = @Schema(implementation = Book.class))
+    ),
+  })
   @PostMapping
   public ResponseEntity<Book> create(@RequestBody Book book) {
     Book createdBook = service.create(book);
     return ResponseEntity.status(HttpStatus.CREATED).body(createdBook);
   }
 
+  @ApiResponses(value = {
+    @ApiResponse(
+      responseCode = "200",
+      description = "Livre mis-à-jour avec succès",
+      content = @Content(schema = @Schema(implementation = Book.class))
+    ),
+    @ApiResponse(
+      responseCode = "404",
+      description = "Livre non-trouvé",
+      content = @Content(schema = @Schema(hidden = true))
+    ),
+    @ApiResponse(
+      responseCode = "400",
+      description = "L'identifiant passé en paramètre et dans l'object Book ne correspondent pas",
+      content = @Content(schema = @Schema(hidden = true))
+    )
+  })
   @PutMapping("{id}")
   public ResponseEntity<Book> update(@PathVariable Long id, @RequestBody Book book) {
     Boolean bookExists = service.exists(id);
 
-    if (!bookExists) ResponseEntity.status(HttpStatus.NOT_FOUND);
-    if (id != book.getId()) ResponseEntity.status(HttpStatus.BAD_REQUEST);
+    if (!bookExists) return ResponseEntity.notFound().build();
+    if (id != book.getId()) return ResponseEntity.badRequest().build();
 
     Book result = service.update(id, book);
     return ResponseEntity.status(HttpStatus.OK).body(result);
   }
 
+  @ApiResponses(value = {
+    @ApiResponse(
+      responseCode = "204",
+      description = "Livre supprimé avec succès",
+      content = @Content(schema = @Schema(hidden = true))
+    ),
+    @ApiResponse(
+      responseCode = "404",
+      description = "Livre non-trouvé",
+      content = @Content(schema = @Schema(hidden = true))
+    ),
+  })
   @DeleteMapping("{id}")
   public ResponseEntity<Void> delete(@PathVariable Long id) {
+    if (!service.exists(id)) return ResponseEntity.notFound().build();
+    
     service.delete(id);
     return ResponseEntity.noContent().build();
   }
